Run honeytail as service user, not root (#40)
This patch creates a `honeycomb` service user in the package preinstall script,
and runs supervised honeytail as that user. We shouldn't need to be root.

Unfortunately this is a bit byzantine for upstart, because AWS Linux uses
an old version of upstart that doesn't support setuid. (At least one user has
previously reported running on AWS Linux). Oh well.

I tested this on Debian stable (systemd) and AWS Linux (upstart).