Add flag to keyval parser to filter incoming mesasge stream (#46)
All other parsers natively filter out garbage in the logs because they all have a strict pattern they're trying to match. The keyval parser treats bare words as keys with the empty string as a value, which means if you have a log line that is *not* in the key=val format, every word appears as a column in the resulting event, and all columns have the empty string as a value. Clearly, this is terrible behavior.
This change adds a flag to the keyval parser to allow you to define a regex, and:
* only lines that match will be parsed (aka filter on a known good field)
or
* only lines that *do not* match will be parsed (aka filter out noise)