When consuming lines, only trim whitespace in parsers (fixes postgres)
This is an issue that currently isn't caught by unit tests. In the postgres
parser, we rely on the existence of leading whitespace to distinguish lines
that are the continuation of a SQL statement, e.g.:

```
2017-11-07 01:43:39 UTC [3542-7] postgres@test LOG:  duration: 15.577 ms  statement: SELECT * FROM test
		WHERE id=1;
```

But in `tail.go`, we're stripping leading and trailing whitespace. Do you
remember what the scenarios were where we needed that? For the moment, I've
been conservative and lifted the `strings.TrimSpace` call into all the parsers
except the postgres one. Not sure if this is the best approach. It does mean
that this won't accidentally break anything.