Make parsing robust to setting NumParsers=0
Currently, if you configure rdslogs to send events directly to Honeycomb, it
will instantiate a MySQL parser, but not set `NumParsers` to a positive value.
So that causes rdslogs to simply hang.

As a general matter of good practice, we might as well make it so that these
parsers work even if `NumParsers` isn't set. Do that in this patch.